+++ src/darray.rs
@@ -1,13 +1,10 @@
 //! Constant-time select data structure over integer sets with the dense array technique by Okanohara and Sadakane.
 #![cfg(target_pointer_width = "64")]
 
-pub mod iter;
-
 use std::io::{Read, Write};
 
 use anyhow::Result;
 
-use crate::darray::iter::Iter;
 use crate::{broadword, BitVector, Searial, Selector};
 
 const BLOCK_LEN: usize = 1024;
@@ -330,6 +327,39 @@ impl DArrayIndex {
 
     fn get_word_over_zero(bv: &BitVector, word_idx: usize) -> usize {
         !bv.words()[word_idx]
+    }
+}
+
+/// Iterator for enumerating integers, created by [`DArray::iter()`].
+pub struct Iter<'a> {
+    da: &'a DArray,
+    pos: usize,
+}
+
+impl<'a> Iter<'a> {
+    /// Creates a new iterator.
+    pub const fn new(da: &'a DArray) -> Self {
+        Self { da, pos: 0 }
+    }
+}
+
+impl<'a> Iterator for Iter<'a> {
+    type Item = usize;
+
+    #[inline(always)]
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.pos < self.da.len() {
+            let x = self.da.select1(self.pos).unwrap();
+            self.pos += 1;
+            Some(x)
+        } else {
+            None
+        }
+    }
+
+    #[inline(always)]
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        (self.da.len(), Some(self.da.len()))
     }
 }
